LayoutTests/imported/w3c/web-platform-tests/html/semantics/interactive-elements/the-dialog-element/dialog-cancel-with-select-expected.txt

CONSOLE MESSAGE: Error: assert_true: select element should be focused expected true got false
Test dialog modal is closed by escape key with select focused



FAIL Test dialog modal is closed by escape key with select focused Error: assert_true: select element should be focused expected true got false
